# web/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCreate": {
        "properties": {
          "batch_size": {
            "default": 16,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "type": "integer"
          },
          "composition_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composition Run Id"
          },
          "dataset_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "dataset_label": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Label"
          },
          "max_bandgap": {
            "default": 1.0,
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Max Bandgap",
            "type": "number"
          },
          "max_e_hull": {
            "default": 0.01,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max E Hull",
            "type": "number"
          },
          "message": {
            "default": "SS-Screen Web control plane",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "min_group_size": {
            "default": 2,
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Min Group Size",
            "type": "integer"
          },
          "min_x_elements": {
            "default": 2,
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Min X Elements",
            "type": "integer"
          },
          "nelems": {
            "default": 2,
            "maximum": 3.0,
            "minimum": 2.0,
            "title": "Nelems",
            "type": "integer"
          },
          "pipeline": {
            "default": "control-plane-demo",
            "enum": [
              "control-plane-demo",
              "stage-1-mp-offline",
              "stage-1a-composition",
              "stage-2-condensation"
            ],
            "title": "Pipeline",
            "type": "string"
          },
          "simulate_failure": {
            "default": false,
            "title": "Simulate Failure",
            "type": "boolean"
          }
        },
        "title": "RunCreate",
        "type": "object"
      },
      "RunResponse": {
        "properties": {
          "configuration": {
            "additionalProperties": true,
            "title": "Configuration",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pipeline_version": {
            "title": "Pipeline Version",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version_counter": {
            "title": "Version Counter",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "pipeline_version",
          "status",
          "configuration",
          "started_at",
          "finished_at",
          "version_counter",
          "created_at",
          "updated_at"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "TaskResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "title": "Depends On",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "scientific_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Status"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version_counter": {
            "title": "Version Counter",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "run_id",
          "stage",
          "task_type",
          "status",
          "scientific_status",
          "depends_on",
          "started_at",
          "finished_at",
          "version_counter",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SS-Screen Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/runs": {
      "get": {
        "operationId": "list_for_project_api_v1_projects__project_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunResponse"
                  },
                  "title": "Response List For Project Api V1 Projects  Project Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List For Project",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_api_v1_projects__project_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "detail_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/actions/cancel": {
      "post": {
        "operationId": "cancel_api_v1_runs__run_id__actions_cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/actions/start": {
      "post": {
        "operationId": "start_api_v1_runs__run_id__actions_start_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}